Modified version of Mattijs' concave tracking modifications.  Changes to Mattijs' work are:
+ Correct use of normals and cosine (dot product) to identify planar
  faces.

+ Correct use for lambda in tracking - it is a fraction, not a distance.

+ Not doing a reduce on demand driven construction data - not all
  processors call it at the same time, so crashes.

This implementation contains an attempt at making the calculation of
lambdaC (from the cell centre) use decomposed triangles for faces, but
this is a bad approach, the concept of using lambdaC relies on the
definition of a convex polyhedron, i.e. none of the planes of the
faces of the polyhedron are inside the volume.  Can't use this method
and will need to treat a convex cell completely differently, not just
some of its faces.