Changing indexedCell to store an integer count of the number of times a dual vertex has been limited.
Allowing the face filtering to continue after warning that mesh
quality cannot be satisfied based on a control Switch.  The filtering
terminates when the number of wrongFaces is less than or equal to the
intial number.

There is a possiblity that the filtering will never converge in this
mode (or in fact, any other).  For example, if a face is on the brink
of being "wrong" and a face motion on the other side of the cell moves
the cell centre enough to make it "wrong", then the limiting mechanism
won't be able to pick this up and stop it, unless limiting is applied
to all points of all cells that are connected to a point that is on a
wrong face.  This can be done if a vertex on a face has been limited
many (say > 4) times.