BUG: sixDoFRigidBodyMotion.  Fixing problems that only occur when orientation != I at start.
Being more careful with when using directions relative to the absolute
axes or the direction of something relative to its initial direction.

Renaming p0, v0 to pInitial, vInitial to be clearer.

fixedOrientation constraint does not need a reference orientation, as
it is not to move at all, relative to any orientation, so the
Cartesian axes are fine.