ENH: removing isSurfacePoint.
It was conceptually broken.  The idea was that in certain cases, you knew that
you were on the surface of the geometry (point pair distance queries for
example) so didn't need to do a nearest-point-on-surface query.  This was fine
for a single surface, but when there were multiple surfaces, the first
high-priority surface would return a size, even if the query point was miles
away.  To fix this would need a distance check, removing the point of the check
altogether.