ENH: cvMesh: Dynamic tree added for surface points.
Now there is a search for the surface points to check whether they are close
to any other surface point. If they are, reject the point unless the angle
between the normals of the surfaces are around 180 degrees, which means that
the points are close but on opposite surfaces (i.e. there is a small gap).