ParaView: Upgrade to 4.3.1
Note that in order to build and install ParaView-4.3.1 lines 653-656
must be removed or commented from file

ThirdParty-dev/ParaView-4.3.1/Qt/Components/CMakeLists.tx

after unpacking the ParaView-4.3.1 source-pack from KitWare.  This is
because the file ui_pqExportStateWizard.h referred to in the lines
653-656:

if(PARAVIEW_INSTALL_DEVELOPMENT_FILES)
  install(FILES "${CMAKE_CURRENT_BINARY_DIR}/ui_pqExportStateWizard.h"
          DESTINATION "${VTK_INSTALL_INCLUDE_DIR}")
endif()

does not exist and cannot be installed causing the build and
installation to fail.