BUG: #1726: was calculating the wrong rotationAngle in decomposePar
In cyclicACMIPolyPatch there is logic to automatically flip the rotation
angle to fit the geometry. However, e.g. when in decomposePar, there is
only part of the geometry (maybe zero faces) and it might make the wrong
decision. Added logic to flip sign only if that makes the error close to 0